feat: add Infrastructure project and update token handling
Updated solution to include new Infrastructure project.
Modified ITokenProvider interface and implemented it in TokenProvider.
Changed Id, PublishDate, and ExpiryDate types in Document and DocumentMetadata.
Added AuthRequest class.
Updated TokenResponse ExpiryDate type.

+++ src/HeadlessCmsClient.Core/Models/TokenResponse.cs
@@ -3,5 +3,5 @@
 public class TokenResponse
 {
     public string BearerToken { get; set; } = string.Empty;
-    public int ExpiryDate { get; set; }
+    public long ExpiryDate { get; set; }
 }
